requirements.txt
flask
flask-cors
yfinance
requests
python-dotenv
pytest
gunicorn
